+++ src/components/Home.jsx
@@ -31,6 +31,7 @@ import View1SSO from "./views/View1SSO";
 import View2DatosIniciales from "./views/View2DatosIniciales";
 import View3DatosDelServicio from "./views/View3DatosDelServicio";
 import View4PeriodoDeServicio from "./views/View4PeriodoDeServicio";
+import View5Herramientas from "./views/View5Herramientas";
 
 moment.locale("es");
 function Home() {
@@ -108,21 +109,6 @@ function Home() {
 
 	const handleBack = () => {
 		setActiveStep((prevActiveStep) => prevActiveStep - 1);
-	};
-
-	const addHerramienta = () => {
-		let arrTemp = [...herramientas];
-		arrTemp.push([calibracion, barcode, herramienta]);
-		setHerramientas(arrTemp);
-		setCalibracion(null);
-		setBarcode("");
-		setHerramienta("");
-	};
-
-	const deleteHerramienta = (i) => {
-		let arrTemp = [...herramientas];
-		arrTemp.splice(i, 1);
-		setHerramientas(arrTemp);
 	};
 
 	const addRefaccion = () => {
@@ -295,85 +281,15 @@ function Home() {
 							/>
 						</div>
 						<div className="views view5">
-							<h3>Herramientas</h3>
-							<div className="herrXagregar">
-								<MuiPickersUtilsProvider utils={MomentUtils}>
-									<DatePicker
-										margin="dense"
-										inputVariant="outlined"
-										autoOk
-										disableToolbar
-										showTodayButton
-										todayLabel="hoy"
-										clearable
-										clearLabel="borrar"
-										okLabel=""
-										cancelLabel=""
-										format="DD/MM/YY"
-										id="calibracion"
-										label="Sig. Calibracion"
-										value={calibracion}
-										onChange={(e) => {
-											setCalibracion(e);
-										}}
-									/>
-								</MuiPickersUtilsProvider>
-								<TextField
-									label="Barcode"
-									fullWidth
-									variant="outlined"
-									size="small"
-									value={barcode}
-									onChange={(e) => setBarcode(e.target.value)}
-								/>
-								<TextField
-									label="Herramienta"
-									fullWidth
-									variant="outlined"
-									size="small"
-									value={herramienta}
-									onChange={(e) => setHerramienta(e.target.value)}
-								/>
-
-								<div className="btnAddHerr">
-									<Button
-										size="small"
-										variant="contained"
-										color="primary"
-										onClick={() => addHerramienta()}>
-										Agregar
-									</Button>
-								</div>
-							</div>
-							<div className="herrAgregadas">
-								{herramientas.length > 0 ? (
-									<>
-										<ul className="ulherr">
-											<li>Sig. Calibración</li>
-											<li>Barcode</li>
-											<li>Herramienta</li>
-											<li>Borrar</li>
-										</ul>
-
-										{herramientas.map((herr, index) => {
-											return (
-												<ul className="ulherr" key={index}>
-													<li>{moment(herr[0]).format("DD/MM/YY")}</li>
-													<li>{herr[1]}</li>
-													<li>{herr[2]}</li>
-													<li>
-														<b
-															className="btnDeleteTime"
-															onClick={() => deleteHerramienta(index)}>
-															X
-														</b>
-													</li>
-												</ul>
-											);
-										})}
-									</>
-								) : null}
-							</div>
+							<View5Herramientas
+								onDone={(_miherramientas) => {
+									setHerramientas(_miherramientas);
+									setDatos({
+										...datos,
+										herramientas: _miherramientas,
+									});
+								}}
+							/>
 						</div>
 						<div className="views view6">
 							<h3>Refacciones</h3>
